test(streaming): isolate public error boundary fixture
